backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
langchain-openai==0.0.5
pyjwt==2.8.0
bcrypt==4.1.1
python-multipart==0.0.6
stripe==7.8.0
httpx==0.25.2
google-auth==2.25.2

